requirements.txt
future
emoji
lxml
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#googletrans
google_trans_new
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.16.4
spamwatch
alphabet_detector
pyrate-limiter
cachetools
fontTools
gtts
pyrogram
redis
nekos.py
tswift
pyowm
tabulate
parsel
pyrogram
tgcrypto
telegraph
pytz
hachoir
pybase64
pySmartDL
validators
youtube_dl
html2text
pyowm==2.8.0